Move loot-core/client/store and loot-core/client/redux to desktop-client package
