+++ packages/desktop-client/src/budgets/budgetsSlice.ts
@@ -1,7 +1,6 @@
 import { createSlice, type PayloadAction } from '@reduxjs/toolkit';
 import { t } from 'i18next';
 
-import { createAppAsyncThunk } from 'loot-core/client/redux';
 import { send } from 'loot-core/platform/client/fetch';
 import { type RemoteFile } from 'loot-core/server/cloud-storage';
 import { getDownloadError, getSyncError } from 'loot-core/shared/errors';
@@ -12,6 +11,7 @@ import { type Handlers } from 'loot-core/types/handlers';
 import { resetApp, setAppState } from '../app/appSlice';
 import { closeModal, pushModal } from '../modals/modalsSlice';
 import { loadGlobalPrefs, loadPrefs } from '../prefs/prefsSlice';
+import { createAppAsyncThunk } from '../redux';
 import { signOut } from '../users/usersSlice';
 
 const sliceName = 'budgets';
