+++ packages/desktop-client/src/gocardless.ts
@@ -1,8 +1,8 @@
-import { type AppDispatch } from 'loot-core/client/store';
 import { send } from 'loot-core/platform/client/fetch';
 import { type GoCardlessToken } from 'loot-core/types/models';
 
 import { pushModal } from './modals/modalsSlice';
+import { type AppDispatch } from './redux';
 
 function _authorize(
   dispatch: AppDispatch,
