+++ packages/desktop-client/src/modals/modalsSlice.ts
@@ -1,6 +1,5 @@
 import { createSlice, type PayloadAction } from '@reduxjs/toolkit';
 
-import { createAppAsyncThunk } from 'loot-core/client/redux';
 import { send } from 'loot-core/platform/client/fetch';
 import { type File } from 'loot-core/types/file';
 import {
@@ -20,6 +19,7 @@ import {
 } from 'loot-core/types/models';
 
 import { resetApp, setAppState } from '../app/appSlice';
+import { createAppAsyncThunk } from '../redux';
 import { signOut } from '../users/usersSlice';
 
 const sliceName = 'modals';
