+++ packages/desktop-client/src/redux/mock.tsx
@@ -1,7 +1,63 @@
 import React, { type ReactNode } from 'react';
 import { Provider } from 'react-redux';
 
-import { mockStore } from 'loot-core/client/store/mock';
+import { configureStore, combineReducers } from '@reduxjs/toolkit';
+
+import {
+  name as accountsSliceName,
+  reducer as accountsSliceReducer,
+} from '../accounts/accountsSlice';
+import {
+  name as appSliceName,
+  reducer as appSliceReducer,
+} from '../app/appSlice';
+import {
+  name as budgetsSliceName,
+  reducer as budgetsSliceReducer,
+} from '../budgets/budgetsSlice';
+import {
+  name as modalsSliceName,
+  reducer as modalsSliceReducer,
+} from '../modals/modalsSlice';
+import {
+  name as notificationsSliceName,
+  reducer as notificationsSliceReducer,
+} from '../notifications/notificationsSlice';
+import {
+  name as prefsSliceName,
+  reducer as prefsSliceReducer,
+} from '../prefs/prefsSlice';
+import {
+  name as queriesSliceName,
+  reducer as queriesSliceReducer,
+} from '../queries/queriesSlice';
+import {
+  name as usersSliceName,
+  reducer as usersSliceReducer,
+} from '../users/usersSlice';
+
+import { type store as realStore } from './index';
+
+const appReducer = combineReducers({
+  [accountsSliceName]: accountsSliceReducer,
+  [appSliceName]: appSliceReducer,
+  [budgetsSliceName]: budgetsSliceReducer,
+  [modalsSliceName]: modalsSliceReducer,
+  [notificationsSliceName]: notificationsSliceReducer,
+  [prefsSliceName]: prefsSliceReducer,
+  [queriesSliceName]: queriesSliceReducer,
+  [usersSliceName]: usersSliceReducer,
+});
+
+export let mockStore: typeof realStore = configureStore({
+  reducer: appReducer,
+});
+
+export function resetMockStore() {
+  mockStore = configureStore({
+    reducer: appReducer,
+  });
+}
 
 export function TestProvider({ children }: { children: ReactNode }) {
   return <Provider store={mockStore}>{children}</Provider>;
