+++ packages/desktop-client/src/setupTests.js
@@ -1,6 +1,5 @@
-import { resetMockStore } from 'loot-core/client/store/mock';
-
 import { installPolyfills } from './polyfills';
+import { resetMockStore } from './redux/mock';
 
 installPolyfills();
