--- packages/loot-core/src/client/redux.ts
@@ -1,8 +1,0 @@
-import { createAsyncThunk } from '@reduxjs/toolkit';
-
-import { type AppDispatch, type RootState } from './store';
-
-export const createAppAsyncThunk = createAsyncThunk.withTypes<{
-  state: RootState;
-  dispatch: AppDispatch;
-}>();
