+++ packages/loot-core/src/client/shared-listeners.ts
@@ -17,13 +17,12 @@ import {
   getCategories,
   getPayees,
 } from '@actual-app/web/src/queries/queriesSlice';
+import { type AppStore } from '@actual-app/web/src/redux';
 import { signOut } from '@actual-app/web/src/users/usersSlice';
 /* eslint-enable no-restricted-imports */
 import { t } from 'i18next';
 
 import { listen, send } from '../platform/client/fetch';
-
-import { type AppStore } from './store';
 
 export function listenForSyncEvent(store: AppStore) {
   // TODO: Should this run on mobile too?
