--- packages/loot-core/src/client/store/index.ts
@@ -1,84 +1,0 @@
-// This is temporary until we move all loot-core/client over to desktop-client.
-/* eslint-disable no-restricted-imports */
-import {
-  name as accountsSliceName,
-  reducer as accountsSliceReducer,
-} from '@actual-app/web/src/accounts/accountsSlice';
-import {
-  name as appSliceName,
-  reducer as appSliceReducer,
-} from '@actual-app/web/src/app/appSlice';
-import {
-  name as budgetsSliceName,
-  reducer as budgetsSliceReducer,
-} from '@actual-app/web/src/budgets/budgetsSlice';
-import {
-  name as modalsSliceName,
-  reducer as modalsSliceReducer,
-} from '@actual-app/web/src/modals/modalsSlice';
-import {
-  name as notificationsSliceName,
-  reducer as notificationsSliceReducer,
-  addNotification,
-} from '@actual-app/web/src/notifications/notificationsSlice';
-import {
-  name as prefsSliceName,
-  reducer as prefsSliceReducer,
-} from '@actual-app/web/src/prefs/prefsSlice';
-import {
-  name as queriesSliceName,
-  reducer as queriesSliceReducer,
-} from '@actual-app/web/src/queries/queriesSlice';
-import {
-  name as usersSliceName,
-  reducer as usersSliceReducer,
-} from '@actual-app/web/src/users/usersSlice';
-/* eslint-enable no-restricted-imports */
-import {
-  combineReducers,
-  configureStore,
-  createListenerMiddleware,
-  isRejected,
-} from '@reduxjs/toolkit';
-
-const rootReducer = combineReducers({
-  [accountsSliceName]: accountsSliceReducer,
-  [appSliceName]: appSliceReducer,
-  [budgetsSliceName]: budgetsSliceReducer,
-  [modalsSliceName]: modalsSliceReducer,
-  [notificationsSliceName]: notificationsSliceReducer,
-  [prefsSliceName]: prefsSliceReducer,
-  [queriesSliceName]: queriesSliceReducer,
-  [usersSliceName]: usersSliceReducer,
-});
-
-const notifyOnRejectedActionsMiddleware = createListenerMiddleware();
-notifyOnRejectedActionsMiddleware.startListening({
-  matcher: isRejected,
-  effect: (action, { dispatch }) => {
-    console.error(action.error);
-    dispatch(
-      addNotification({
-        notification: {
-          id: action.type,
-          type: 'error',
-          message: action.error.message || 'An unexpected error occurred.',
-        },
-      }),
-    );
-  },
-});
-
-export const store = configureStore({
-  reducer: rootReducer,
-  middleware: getDefaultMiddleware =>
-    getDefaultMiddleware({
-      // TODO: Fix this in a separate PR. Remove non-serializable states in the store.
-      serializableCheck: false,
-    }).prepend(notifyOnRejectedActionsMiddleware.middleware),
-});
-
-export type AppStore = typeof store;
-export type RootState = ReturnType<typeof store.getState>;
-export type AppDispatch = typeof store.dispatch;
-export type GetRootState = typeof store.getState;
