+++ upcoming-release-notes/4827.md
@@ -1,0 +1,6 @@
+---
+category: Maintenance
+authors: [joel-jeremy]
+---
+
+Move loot-core/client/store and loot-core/client/redux to desktop-client package
